Fix bug in removing entities from spatial hash
When removing an entity from the spatial hash, component counts would
decrement once when the position component was removed, and one when the
counted component was removed. This was because the spatial hash was
using the current state of the ecs to determine whether an entity has a
particular component, when it should have been using the current ecs
state plus the action being applied.