Panic when expected component is missing
Previously these would result in errors that propogate up and are
handled by the game launcher. This led to difficult-to-debug missing
component errors, since to information was provided about the point in
the program when a component was expected but missing. Now, whenever a
component is expected but missing the program panics with a usable error
message.